ref(group_index): move get_group_list to group_index/lookup.py and export from __init__
get_group_list is a read-only lookup helper with no mutations — it doesn't
belong in the heavy update.py module. Moving it to a minimal lookup.py
avoids a circular-import risk (update.py ← index.py ← __init__.py ←
update.py) and gives the function a permanent, officially-exported home via
group_index.__init__.__all__.

Callers should now import from sentry.api.helpers.group_index directly.
