+++ src/sentry/api/helpers/group_index/__init__.py
@@ -24,11 +24,13 @@ __all__ = (
     "BULK_MUTATION_LIMIT",
     "SEARCH_MAX_HITS",
     "delete_group_list",
+    "get_group_list",
     "update_groups",
 )
 
 from .delete import *  # NOQA
 from .delete import delete_group_list
 from .index import *  # NOQA
+from .lookup import get_group_list
 from .update import *  # NOQA
 from .update import update_groups
